python/proxies-weather/requirements.txt
python-dotenv
pydantic
stagehand

